CL-44 Done, ready for review. Barmanager can delete accounts, but can not delete their own account or edit their own role. This is secured in front- and back-end. Added thymeleaf variable to show/hide delete/role options. Deleted border from /products.

+++ src/main/java/com/MIW/Cohort5/Clups/services/implementations/UserDetailsServiceImpl.java
@@ -9,6 +9,8 @@ import com.MIW.Cohort5.Clups.services.RoleService;
 import com.MIW.Cohort5.Clups.services.UserService;
 import com.MIW.Cohort5.Clups.services.dtoConverters.UserDtoConverter;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -32,6 +34,7 @@ public class UserDetailsServiceImpl implements UserService {
     private static final String INITIAL_PASSWORD_ROOT = "pw";
     private static final String INITIAL_USERNAME_ROOT = "user";
     private static final String STANDARD_ROLE = "CUSTOMER";
+    private static final String MANAGER_ROLE = "BARMANAGER";
 
     private UserDtoConverter dtoConverter = new UserDtoConverter();
 
@@ -101,6 +104,10 @@ public class UserDetailsServiceImpl implements UserService {
     }
 
     private UserDto editRole(UserDto userDto) {
+
+        if (loggedInUser(userDto)) {
+            throw new IllegalArgumentException("You are not allowed to change your own role");
+        }
 
         if (userDto.getUserRole() == null) {
             userDto.setUserRole(userRepository.findUserByUserCode(userDto.getUserCode()).getRole().getRoleName());
@@ -222,9 +229,14 @@ public class UserDetailsServiceImpl implements UserService {
 
     @Override
     public User deleteUser(UserDto userDto){
-        User model = findModelByUserCode(userDto.getUserCode());
-        userRepository.delete(model);
-        return model;
+
+        if (loggedInUser(userDto)) {
+            throw new IllegalArgumentException("You are not allowed to remove your own account");
+        } else {
+            User model = findModelByUserCode(userDto.getUserCode());
+            userRepository.delete(model);
+            return model;
+        }
     }
 
     public void addCredit(Integer userCode, Integer amount) {
@@ -237,6 +249,17 @@ public class UserDetailsServiceImpl implements UserService {
 
     public List<Integer> getUserByPartialString(String request) {
         return userRepository.findUserByPartialString(request);
+    }
+
+    @Override
+    public boolean loggedInUser(UserDto userDto) {
+        Authentication loggedInUser = SecurityContextHolder.getContext().getAuthentication();
+
+        if (userDto != null) {
+            return loggedInUser.getName().equals(userDto.getUsername());
+        } else {
+            return false;
+        }
     }
 
 }
